[kt, tests] Rewrite tests to have the one single Gradle build
Previously, testHmppLibAndMppConsumer and testMppLibAndHmppConsumer had
two different Gradle builds (two different independent settings.gradle.kts)

This is more realistic setup, but it leads to some annoying issues in the
tests infra (e.g., it's harder to force JDK for Gradle VM for all projects,
so the 'lib'-project was using default JDK home and failing due to AGP
requirements).

This commit replaces that setup with one single Gradle build (one
settings.gradle.kts), which includes 'lib'-project ('include("lib")'), but doesn't
depend on it as sources, only on binaries. This is a bit more esoteric, but
but we were not able to fund why it's bad for the purposes of testing

(cherry picked from commit 13622444ece0de98d1c99dd98885d1a0d89a4283)

GitOrigin-RevId: d9440875c2df222bfcc5db518eab3855ed8e736b